Rename test_tlog to rec and move helper functions
Moving test_tlog.py to test_tlog_rec.py to prepare for adding new
test cases for other tlog commands.

Moving helper functions out of test module and into a separate
misc module
